Update syslog server, continue refactor, allow testing without instrumentation

+++ discover_test_coverage/file.py
@@ -19,6 +19,34 @@ def find_python_files(directory: Path) -> List[Path]:
     return all_python_files
 
 
+def find_python_files_not_conftest(directory: Path) -> List[Path]:
+    """Find all of the Python files that are not a conftest.py file."""
+    # create a list to store Python files that are not conftest.py
+    not_conftest_file_list = []
+    # find all of the Python files in the specified directory
+    python_file_list = find_python_files(directory)
+    # add to the not_conftest_file_list all files that
+    # are not conftest.py according to their name
+    for python_file in python_file_list:
+        if python_file.name != constants.tests.Conftest:
+            not_conftest_file_list.append(python_file)
+    return not_conftest_file_list
+
+
+def find_conftest_files(directory: Path) -> List[Path]:
+    """Find all of the Python files that are a conftest.py file."""
+    # create a list of all the conftest.py files
+    conftest_file_list = []
+    # find all of the Python files in the specified directory
+    python_file_list = find_python_files(directory)
+    # add to the conftest_file_list all files that are
+    # conftest.py according to their name
+    for python_file in python_file_list:
+        if python_file.name == constants.tests.Conftest:
+            conftest_file_list.append(python_file)
+    return conftest_file_list
+
+
 def elide_path(path, maximum_parts: int = 4, include_up_to: int = 5):
     """Elide a path so that it supports compact display."""
     parts = list(PurePath(path).parts)
@@ -31,7 +59,9 @@ def elide_path(path, maximum_parts: int = 4, include_up_to: int = 5):
     return PurePath(*parts)
 
 
-def create_hidden_directory(containing_directory: Path, directory: Path) -> Path:
+def create_hidden_directory(
+    containing_directory: Path, directory: Path, delete_path: bool = True
+) -> Path:
     """Create a hidden directory."""
     output.logger.debug(f"Hiding in containing directory: {containing_directory}")
     output.logger.debug(f"Hiding the directory with name: {directory.name}")
@@ -41,10 +71,27 @@ def create_hidden_directory(containing_directory: Path, directory: Path) -> Path
         containing_directory / Path(constants.markers.Hidden + str(directory.name))
     )
     # delete the hidden directory if it already exists
-    rmtree(hidden_directory)
+    if hidden_directory.exists() and hidden_directory.is_dir() and delete_path:
+        rmtree(hidden_directory)
     # create the hidden directory for storing instrumentation
     # note that mkdir throws an exception if the directory exists;
     # since it was already deleted, the exception should not occur
     hidden_directory.mkdir(parents=True)
     output.logger.debug(f"Created the hidden directory: {hidden_directory}")
+    return hidden_directory
+
+
+def get_hidden_directory(
+    containing_directory: Path,
+    directory: Path,
+) -> Path:
+    """Create a hidden directory."""
+    output.logger.debug(f"Hiding in containing directory: {containing_directory}")
+    output.logger.debug(f"Hiding the directory with name: {directory.name}")
+    # access a hidden directory (i.e., a directory with name "program" will
+    # lead to creation of a hidden directory called ".program")
+    hidden_directory = Path(
+        containing_directory / Path(constants.markers.Hidden + str(directory.name))
+    )
+    output.logger.debug(f"Accessing the hidden directory: {hidden_directory}")
     return hidden_directory
